requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
idna==3.3
python-dateutil==2.8.2
requests==2.27.1
responses==0.20.0
six==1.16.0
urllib3==1.26.9
